refactor: move `VALIDATION_ERROR_FOOTER ` to `constants`
